[REFACTOR] 디렉토리명 수정 error -> exception
